+++ flutter_inappwebview_forge_android/android/src/main/kotlin/com/emirkanacar/flutter_inappwebview_forge_android/plugin_scripts_js/JavaScriptBridgeJS.kt
@@ -354,6 +354,7 @@ class JavaScriptBridgeJS private constructor() {
             "  })(window);" +
             "}" +
             "if (window." + get_JAVASCRIPT_BRIDGE_NAME() + " != null) {" +
+            "  " + WebMessageListenerJS.WEB_MESSAGE_LISTENER_JS_SOURCE() +
             "  " + UTIL_JS_SOURCE() +
             "}";
   }